[Rename] RegisteredTags→Tagsに変更 長くて見づらいので
